Centralized resource table generation code
