+++ dhis-2/dhis-services/dhis-service-dxf2/src/test/java/org/hisp/dhis/dxf2/metadata/objectbundle/ObjectBundleServiceTest.java
@@ -1371,7 +1371,6 @@ public class ObjectBundleServiceTest
 
         ObjectBundle bundle = objectBundleService.create( params );
         ObjectBundleValidationReport validate = objectBundleValidationService.validate( bundle );
-        System.err.println( "V: " + validate.getErrorReports() );
         assertTrue( validate.getErrorReports().isEmpty() );
 
         objectBundleService.commit( bundle );
